Add DHCP, ping, DMA and e1000 improvements
Introduce a DHCP client and ping tooling, add a thin DMA API, and improve e1000 driver reliability and integration. Files added: DHCP client (src/net/dhcp.{c,h}), DMA wrapper (src/mem/dma.{c,h}), and apps/apps.ld. Networking changes: UDP now dispatches DHCP replies, net core and loopback use 32-bit packet lengths, IP id is now a monotonically-incremented value, and net_init attempts DHCP on the primary NIC. e1000 changes: use dma_alloc_page for descriptors/buffers, track TX drops, handle 64-bit PCI BARs, poll until reset clears, reprogram RAR0 after reset, and remove promiscuous flags. Shell: add ping command, parser and ICMP echo logic; shell links required net headers. Memory fixes: PMM bit ops use unsigned masks, heap free adds backward coalescing and replaces an early return with an infinite panic loop on non-contiguous heap pages. Video/kernel: terminal_clear now repaints framebuffer and resets cursor to 0,0; kpanic writes directly to real framebuffer to ensure BSOD is visible. Minor changes: .gitignore tweaks (.DS_Store), timezone adjustment in settings, and small formatting/cleanup across headers and initializations.

+++ src/net/net.h
@@ -78,7 +78,7 @@ bool net_has_primary_nic(void);
 void net_init(void);
 
 // Core receive entrypoint: hardware drivers should call this with an Ethernet frame.
-void net_handle_packet(void* packet, uint16_t packet_length, net_nic_interfaces_t* nic);
+void net_handle_packet(void* packet, uint32_t packet_length, net_nic_interfaces_t* nic);
 
 // Bring-up debug helpers
 uint64_t net_dbg_get_rx_frames(void);
@@ -87,6 +87,6 @@ void net_dbg_reset(void);
 
 // Optional helper: deliver a packet in a task context. If you don't have a thread
 // system like Polaris, you can just call net_handle_packet directly.
-void net_handle_packet_deferred(void* packet, uint16_t packet_length, net_nic_interfaces_t* nic);
+void net_handle_packet_deferred(void* packet, uint32_t packet_length, net_nic_interfaces_t* nic);
 
 #endif
